1_모델링/PAN/requirements.txt
Bottleneck==1.3.2
certifi==2021.10.8
click==7.1.2
MarkupSafe==2.0.1
mkl-fft==1.3.1
mkl-service==2.4.0
munkres==1.1.4
opencv-python==4.5.5.62
pathtools==0.1.2
Pillow==8.4.0
protobuf==3.19.1
pywin32==303
termcolor==1.1.0
torch==1.9.0+cu111
torchinfo==1.6.3
torchvision==0.10.0+cu111
wandb==0.12.10
Werkzeug==2.0.3
wincertstore==0.2
numpy==1.21.2
pandas==1.3.5
matplotlib==3.5.0
